+++ packages/wonder-blocks-icon-button/src/util/icon-button-util.ts
@@ -12,6 +12,8 @@ export const iconSizeForButtonSize = (size: IconButtonSize): IconSize => {
             return "medium";
         case "medium":
             return "medium";
+        case "large":
+            return "medium";
     }
 };
 
@@ -19,4 +21,4 @@ export const iconSizeForButtonSize = (size: IconButtonSize): IconSize => {
  * A function that returns the size of the touch target in pixels for a given icon button size.
  */
 export const targetPixelsForSize = (size: IconButtonSize): number =>
-    ({xsmall: 24, small: 32, medium: 40}[size]);
+    ({xsmall: 24, small: 32, medium: 40, large: 48}[size]);